requirements.txt
asgiref==3.5.2
attrs==22.1.0
boto3==1.24.73
botocore==1.27.73
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
grpcio==1.48.1
grpcio-testing==1.48.1
grpcio-tools==1.48.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
minio==7.1.11
mongomock==4.1.2
moto==4.0.3
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
responses==0.21.0
s3transfer==0.6.0
sentinels==1.0.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
xmltodict==0.13.0
